requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
asyncpg==0.30.0
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.4
dill==0.4.0
distro==1.9.0
fastapi==0.115.0
flake8==7.3.0
Flask==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
mccabe==0.7.0
mistune==3.1.4
mypy_extensions==1.1.0
openai==1.100.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==6.2.1
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.36
structlog==25.4.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn[standard]==0.32.0
websockets==13.1
